Move tests to tests/ folder
