+++ src/tests/App.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import App from './App'
+import App from '../App'
 import { BrowserRouter } from 'react-router-dom'
 import ReactDOM from 'react-dom'
